[fix] showConfirm 로 알림창 변경

+++ src/components/app/proEnter/ProEnterTab3.jsx
@@ -52,11 +52,16 @@ const ProEnterTab3 = () => {
 
             const res = await proEnterSaveData.mutateAsync(payload);
             if (res?.success === "777") {
-                modal.showAlert("알림", "프로젝트가 등록되었습니다.");
-                sessionStorage.setItem("projectnum", "");
-                sessionStorage.setItem("projectname", "");
-                sessionStorage.setItem("servername", "");
-                navigate("/"); //프로젝트 목록 페이지로 이동
+                modal.showConfirm("알림", "프로젝트가 등록되었습니다.", {
+                    btns: [{
+                        title: "확인", click: () => {
+                            sessionStorage.setItem("projectnum", "");
+                            sessionStorage.setItem("projectname", "");
+                            sessionStorage.setItem("servername", "");
+                            navigate("/"); //프로젝트 목록 페이지로 이동
+                        }
+                    }],   ////문항 목록 페이지로 이동
+                });
             } else if (res?.success === "765") {
                 // 중복
                 modal.showErrorAlert("알림", "이미 등록된 프로젝트 입니다.");
